+++ resources/assets/js/components/people/calls/PhoneCallList.vue
@@ -137,7 +137,7 @@
             <form-textarea
               v-model="editCall.content"
               :label="$t('people.modal_call_comment')"
-              rows="4"
+              :rows="4"
               iclass="br2 f5 w-100 ba b--black-40 pa2 outline-0"
               @contentChange="updateEditCallContent($event)"
             />