+++ CHANGELOG.md
@@ -14,6 +14,9 @@
 
 ### Fixes:
 
+* Fix conversation creation
+* Fix phone call update
+* Fix conversation list show
 * Fix subscription cancel
 * Fix last consulted contact list
 * Fix exception in case a user register twice